requirements.txt
discord.py==2.6.3
python-dotenv
espn-api
openai
httpx
